+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.15.1-beta.3] - 2025-01-23
+
+- remove unrelated type params of `QuadTree`
+- support center setting
+
 ## [0.15.1-beta.2] - 2025-01-23
 
 - improved the organization of public exports (`pub use`) to enhance usability and clarity.